+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Add support for `tuned.openshift.io` apiGroup in OpenShiftClient DSL
 * Add DSL support for ConsolePlugin and ConsoleQuickStart in `console.openshift.io` apiGroup
 * Add DSL support for `user.openshift.io/v1` Identity in OpenShiftClient DSL
+* Add DSL support for OpenShift Whereabouts CNI Model `whereabouts.cni.cncf.io` to OpenShiftClient DSL
 
 #### _**Note**_: Breaking changes in the API
 ##### DSL Changes: